Default Implementation in Rule integrated
